+++ Providers/IntercomProvider.h
@@ -1,0 +1,7 @@
+#import "ARAnalyticalProvider.h"
+
+@interface IntercomProvider : ARAnalyticalProvider
+
+- (instancetype)initWithWithAppID:(NSString *)identifier apiKey:(NSString *)apiKey;
+
+@end