+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 3.3.0
 
 * Added support for extra properties on screen views ( @alloy )
+* Corrected interface of `+[ARAnalytics setupFabricWithKits:]`. (from NSDictionary to NSArray) ( @sodastsai )
 
 ## Version 3.2.0