+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ARAnalytics
 
+## Version 3.4.0
+
+* Added support for Intercom ( @orta )
+
 ## Version 3.3.0
 
 * Added support for extra properties on screen views ( @alloy )